refac: #8 split arr utils
- Split arr utils each in its own file
