+++ src/arr/fill.ts
@@ -1,0 +1,7 @@
+import { from } from "./from";
+
+const fill = <T>(n: number, val: T): T[] => {
+  return from(n, () => val);
+};
+
+export { fill };
