+++ src/arr/range.ts
@@ -1,0 +1,10 @@
+import { from } from "./from";
+
+const range = (a: number, b: number, step = 1): number[] => {
+  const n = Math.ceil((b - a) / step);
+  if (n <= 0) {
+    throw Error("Invalid range");
+  }
+  return from(n, (i) => a + step * i);
+};
+export { range };
